+++ src/web-socket/web-socket.constants.ts
@@ -1,0 +1,8 @@
+export const WSS_PORT = 8080
+
+export enum WsActions {
+    mouseUp = 'mouse_up',
+    mouseDown = 'mouse_down',
+    mouseLeft = 'mouse_left',
+    mouseRight = 'mouse_right',
+}
